advanced mysql

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -1,7 +1,7 @@
 -- safe divide function to avoid 0 divide errors
 DROP FUNCTION IF EXISTS SafeDiv;
 DELIMITER $$
-CREATE FUNCTION SafeDiv(a INT, b INT) RETURNS INT
+CREATE FUNCTION SafeDiv(a INT, b INT) RETURNS FLOAT
 BEGIN
 IF b = 0 THEN
     RETURN 0;
